requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pyyaml==6.0.1
neo4j==5.16.0
google-genai==1.0.0
pydantic==2.5.3
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
